fix(avatarEditing): Avatarchanges with ranomizeButton were not generated correctly
When the user clicked the locks in the avatar editor and then clicked the randomize button, the avatar components that were locked, still generated a new image. The reason for that was the wrong implementation of that each if in the randomize method set the imagecode instead of only one time at the end. Now it is fixed.